Reduce compilation time by re-arranging xcode definitions
Compilation of mdservice/md_fops_xc.h by GCCXML2XC or CASTXML2XC takes a long
time, because mdservice/md_fops.h includes a lot of system and motr headers.

Create a dedicated fop/wire.h with little more than xcode definitions, that can
be compiled faster. Adjust #include-s.

$ time make -j32 # from scratch

Without the patch:

real	2m52.460s
user	7m10.125s
sys	1m2.583s

With the patch:

real	1m23.514s
user	5m56.185s
sys	1m4.003s
