+++ fop/Kbuild.sub
@@ -1,9 +1,7 @@
 m0tr_objects += fop/fop.o \
                   fop/fop_item_type.o \
-                  fop/fop_xc.o \
                   fop/fom.o \
                   fop/fom_generic.o \
                   fop/fom_simple.o \
-                  fop/fom_generic_xc.o \
                   fop/fom_long_lock.o \
                   fop/fom_interpose.o
